Create 1758-07-12-ToEbenezerBlackwell
There is not a specified persID for William Legge so I left comments by these points.
I believe I found an error in the third footnote (71st on page 129 in PDF) where the date for the letter mentioned is incorrectly put for the year 1756 instead of 1758. I put a comment there as well.