Create pers1213
Unsure on pers ID for John Johnson and 1162 bio did not offer much clarification.  I chose this ID over 1163 as the age was similar to Dorothea's even though it is not uncommon for husbands to be far older, just a temporary educational guess until further information proves otherwise. Comments left in listperson line 31 and by mention of name in line 61